Cleanup Metadata complexity
