+++ packages/type-extrinsics/src/fromMetadata.spec.ts
@@ -3,7 +3,7 @@
 // of the Apache-2.0 license. See the LICENSE file for details.
 
 import Metadata from '@polkadot/types/Metadata';
-import json from '@polkadot/types/Metadata.rpc';
+import json from '@polkadot/types/Metadata/static';
 
 import fromMetadata from './fromMetadata';
