+++ packages/type-extrinsics/src/utils/createUnchecked.ts
@@ -2,7 +2,7 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import { FunctionMetadata } from '@polkadot/types/Metadata';
+import { FunctionMetadata } from '@polkadot/types/Metadata/Modules';
 import { ExtrinsicFunction } from '@polkadot/types/Method';
 import { Extrinsic, Method } from '@polkadot/types/index';
 import { assert } from '@polkadot/util';
