+++ packages/type-storage/src/static.ts
@@ -2,7 +2,7 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import metadataRpc from '@polkadot/types/Metadata.rpc';
+import metadataRpc from '@polkadot/types/Metadata/static';
 import Metadata from '@polkadot/types/Metadata';
 
 import fromMetadata from './fromMetadata';
