+++ packages/types/src/EventRecord.spec.ts
@@ -10,7 +10,7 @@ import AccountId from './AccountId';
 import Event from './Event';
 import EventRecord from './EventRecord';
 import Metadata from './Metadata';
-import metadataRpc from './/Metadata.rpc';
+import metadataRpc from './Metadata/static';
 
 describe('EventRecord', () => {
   beforeEach(() => {
