+++ packages/types/src/Extrinsic.ts
@@ -13,7 +13,7 @@ import Struct from './codec/Struct';
 import Address from './Address';
 import ExtrinsicSignature from './ExtrinsicSignature';
 import Hash from './Hash';
-import { FunctionMetadata } from './Metadata';
+import { FunctionMetadata } from './Metadata/Modules';
 import Method from './Method';
 
 type ExtrinsicValue = {
