+++ packages/types/src/Metadata/static.ts
