+++ packages/types/src/StorageKey.ts
@@ -7,7 +7,7 @@ import { AnyU8a } from './types';
 import { isFunction } from '@polkadot/util';
 
 import Bytes from './Bytes';
-import { StorageFunctionMetadata } from './Metadata';
+import { StorageFunctionMetadata } from './Metadata/Modules';
 
 export interface StorageFunction {
   (arg?: any): Uint8Array;
