+++ packages/types/src/scripts/MetadataJsonDecode.ts
@@ -5,7 +5,7 @@
 const fs = require('fs');
 
 import Metadata from '../Metadata';
-import rpcdata from '../Metadata.rpc';
+import rpcdata from '../Metadata/static';
 
 function metadataStringifyAsJson () {
   const decoded = new Metadata(rpcdata);
