+++ packages/types/src/scripts/MetadataMd.ts
@@ -7,7 +7,7 @@ import { stringCamelCase, stringLowerFirst } from '@polkadot/util';
 
 import interfaces from '../../../type-jsonrpc/src';
 import Metadata from '../Metadata';
-import rpcdata from '../Metadata.rpc';
+import rpcdata from '../Metadata/static';
 import Method from '../Method';
 
 const ANCHOR_TOP = `<a id='top' style='text-decoration: none;'>`;
